applications/legacy-admin/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.9
gunicorn==20.1.0
